Move fixture in different file.
